requirements.txt
django>=4.2,<5.0
python-binance>=1.0.19
python-dotenv>=0.20
celery>=5.2.0,<6.0
redis>=4.0.0,<5.0
cryptography>=42.0.0,<43.0.0
requests>=2.28.0,<3.0.0
pandas-ta
numpy==1.26.4
newsapi-python>=0.2.7
